bot/requirements.txt
aiogram==3.4.1
asyncpg==0.31.0
python-dotenv==1.0.1
yookassa==3.3.0
